Refactored volume migration suite
