+++ pkg/cmd/testcmd.go
@@ -18,6 +18,7 @@ package cmd
 
 import (
 	"fmt"
+	volumemigration "github.com/dell/cert-csi/pkg/testcore/suites/migration"
 	"time"
 
 	"github.com/dell/cert-csi/pkg/plotter"
@@ -309,7 +310,7 @@ func getVolumeMigrateCommand(globalFlags []cli.Flag) cli.Command {
 				return fmt.Errorf("failed to get test image: %s", err)
 			}
 			s := []suites.Interface{
-				&suites.VolumeMigrateSuite{
+				&volumemigration.VolumeMigrateSuite{
 					TargetSC:     targetSC,
 					VolumeNumber: volNum,
 					PodNumber:    podNum,
