+++ pkg/testcore/suites/common/common.go
@@ -1,0 +1,24 @@
+package common
+
+import "github.com/dell/cert-csi/pkg/observer"
+
+func GetAllObservers(obsType observer.Type) []observer.Interface {
+	if obsType == observer.EVENT {
+		return []observer.Interface{
+			&observer.PvcObserver{},
+			&observer.VaObserver{},
+			&observer.PodObserver{},
+			&observer.EntityNumberObserver{},
+			&observer.ContainerMetricsObserver{},
+		}
+	} else if obsType == observer.LIST {
+		return []observer.Interface{
+			&observer.PvcListObserver{},
+			&observer.VaListObserver{},
+			&observer.PodListObserver{},
+			&observer.EntityNumberObserver{},
+			&observer.ContainerMetricsObserver{},
+		}
+	}
+	return []observer.Interface{}
+}
